fixed all warnings produced by Visual Studio 2012

+++ python/omegacomplete/core/Algorithm.cpp
@@ -4,8 +4,8 @@
 #include "LookupTable.hpp"
 #include "CompletionPriorities.hpp"
 
-static const size_t kWordSizeCutoffPointForDepthLists = 5;
-static const size_t kMaxDepthPerIndex = 3;
+static const unsigned kWordSizeCutoffPointForDepthLists = 5;
+static const unsigned kMaxDepthPerIndex = 3;
 boost::unordered_map<size_t, std::vector<std::vector<size_t> > >
     Algorithm::depth_list_cache_;
 
@@ -18,7 +18,7 @@ boost::mutex Algorithm::title_case_mutex_;
 void Algorithm::InitStatic()
 {
     // generate depth lists
-    for (size_t ii = 1; ii < kWordSizeCutoffPointForDepthLists; ++ii) {
+    for (unsigned ii = 1; ii < kWordSizeCutoffPointForDepthLists; ++ii) {
         Algorithm::GenerateDepths(ii, kMaxDepthPerIndex, depth_list_cache_[ii]);
 
         //foreach (std::vector<size_t> line, depth_list_cache_[ii]) {
